Remove the anonymous services again, they do not work
Apparently, calling launch will create a new instance of the class.
Thus, there is no way to inject test data into the application. The only
way to add code that acts depending on the state of the greetings
services list... i.e. if it is null add some mock services. Sad.